requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
absl-py==1.0.0
    # via tensorboard
anyio==3.5.0
    # via starlette
asgiref==3.5.0
    # via uvicorn
bcrypt==3.2.0
    # via -r requirements.in
beautifulsoup4==4.11.1
    # via gdown
cachetools==5.0.0
    # via google-auth
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via bcrypt
charset-normalizer==2.0.11
    # via requests
click==8.0.3
    # via
    #   nltk
    #   uvicorn
fastapi==0.73.0
    # via -r requirements.in
filelock==3.6.0
    # via gdown
gdown==4.4.0
    # via -r requirements.in
google-auth==2.6.0
    # via
    #   google-auth-oauthlib
    #   tensorboard
google-auth-oauthlib==0.4.6
    # via tensorboard
grpcio==1.43.0
    # via tensorboard
gunicorn==20.1.0
    # via -r requirements.in
h11==0.13.0
    # via uvicorn
idna==3.3
    # via
    #   anyio
    #   requests
importlib-metadata==4.10.1
    # via markdown
joblib==1.1.0
    # via nltk
markdown==3.3.6
    # via tensorboard
nltk==3.7
    # via -r requirements.in
numpy==1.21
    # via
    #   scipy
    #   tensorboard
oauthlib==3.2.0
    # via requests-oauthlib
passlib==1.7.4
    # via -r requirements.in
protobuf==3.19.4
    # via tensorboard
pyasn1==0.4.8
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.2.8
    # via google-auth
pycparser==2.21
    # via cffi
pydantic==1.9.0
    # via fastapi
pyjwt==2.3.0
    # via -r requirements.in
pysocks==1.7.1
    # via requests
python-dotenv==0.19.2
    # via -r requirements.in
python-multipart==0.0.5
    # via -r requirements.in
regex==2022.3.15
    # via nltk
requests[socks]==2.27.1
    # via
    #   gdown
    #   requests-oauthlib
    #   tensorboard
requests-oauthlib==1.3.1
    # via google-auth-oauthlib
rsa==4.8
    # via google-auth
scipy==1.8.0
    # via -r requirements.in
six==1.16.0
    # via
    #   absl-py
    #   bcrypt
    #   gdown
    #   google-auth
    #   grpcio
    #   python-multipart
sniffio==1.2.0
    # via anyio
soupsieve==2.3.2.post1
    # via beautifulsoup4
starlette==0.17.1
    # via fastapi
tensorboard==2.8.0
    # via -r requirements.in
tensorboard-data-server==0.6.1
    # via tensorboard
tensorboard-plugin-wit==1.8.1
    # via tensorboard
torch==1.10.2
    # via -r requirements.in
tqdm==4.62.3
    # via
    #   -r requirements.in
    #   gdown
    #   nltk
typing-extensions==4.0.1
    # via
    #   pydantic
    #   torch
urllib3==1.26.8
    # via requests
uvicorn==0.17.1
    # via -r requirements.in
werkzeug==2.0.2
    # via tensorboard
wheel==0.37.1
    # via tensorboard
zipp==3.7.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
